refactor(circuit): split sawtooth physics out of circuit_utils [Clean Code: SRP]
Last slice of the B1 split. Move to a new `scqubits/core/sawtooth.py`:

- `sawtooth_potential`
- `sawtooth_operator`

Both are public (no leading `_`) and describe a published junction
model (the sawtooth current-phase relationship), so a re-export shim
in `circuit_utils.py` stays. Internal callers (`circuit_routines.py`,
`circuit_plotting.py`, `tests/test_circuit_utils.py`) now import from
`sawtooth` directly.

Also reorder docstring placement so `sawtooth_potential` is defined
before `sawtooth_operator` (which calls it) in the new module — the
old ordering was reversed and only worked because of name resolution
at call-time.

This concludes B1. `circuit_utils.py` is now down to ~85 lines holding
only the orphan helpers (`truncation_template`, `get_trailing_number`,
`example_circuit`) plus the public deprecation shims.
